+++ packages/esm-patient-common-lib/src/orders/index.ts
@@ -2,3 +2,4 @@ export * from './useOrderBasket';
 export * from './postOrders';
 export * from './useOrders';
 export * from './types';
+export * from './useOrderableConceptSets';
