+++ packages/esm-patient-tests-app/src/test-orders/add-test-order/test-order.ts
@@ -1,5 +1,5 @@
-import { type TestOrderBasketItem, type OrderUrgency } from '@openmrs/esm-patient-common-lib';
-import { type OrderableConcept } from './useOrderableConceptSets';
+import { type TestOrderBasketItem, type OrderUrgency, type OrderableConcept } from '@openmrs/esm-patient-common-lib';
+import { type TestType } from './useTestTypes';
 
 type LabOrderRequest = Pick<TestOrderBasketItem, 'action' | 'testType'>;
 
@@ -15,7 +15,7 @@ export const priorityOptions: PriorityOption[] = [
 ];
 // TODO add priority option `{ value: "ON_SCHEDULED_DATE", label: "On scheduled date" }` once the form supports a date.
 
-export function createEmptyLabOrder(testType: OrderableConcept, orderer: string): TestOrderBasketItem {
+export function createEmptyLabOrder(testType: TestType, orderer: string): TestOrderBasketItem {
   return {
     action: 'NEW',
     urgency: priorityOptions[0].value as OrderUrgency,
