+++ packages/esm-patient-tests-app/src/test-orders/add-test-order/useTestTypes.test.ts
@@ -3,7 +3,7 @@ import useSWRImmutable from 'swr/immutable';
 import { renderHook, waitFor } from '@testing-library/react';
 import { getDefaultsFromConfigSchema, openmrsFetch, restBaseUrl, useConfig } from '@openmrs/esm-framework';
 import { type ConfigObject, configSchema } from '../../config-schema';
-import { useOrderableConcepts } from './useOrderableConceptSets';
+import { useTestTypes } from './useTestTypes';
 
 jest.mock('swr/immutable');
 
@@ -41,17 +41,17 @@ mockOpenrsFetch.mockImplementation((url: string) => {
 
 describe('useTestTypes is configurable', () => {
   it('should return all test concepts when no labOrderableConcepts are provided', async () => {
-    const { result } = renderHook(() => useOrderableConcepts('', '', []));
+    const { result } = renderHook(() => useTestTypes('', [], []));
     expect(mockOpenrsFetch).toHaveBeenCalledWith(
       `${restBaseUrl}/concept?class=Test?v=custom:(display,names:(display),uuid,setMembers:(display,uuid,names:(display),setMembers:(display,uuid,names:(display))))`,
     );
     await waitFor(() => expect(result.current.isLoading).toBeFalsy());
     expect(result.current.error).toBeFalsy();
-    expect(result.current.orderableConcepts).toEqual([expect.objectContaining({ label: 'Test concept' })]);
+    expect(result.current.testTypes).toEqual([expect.objectContaining({ label: 'Test concept' })]);
   });
 
   it('should return children of labOrderableConcepts when provided', async () => {
-    const { result } = renderHook(() => useOrderableConcepts('', '', []));
+    const { result } = renderHook(() => useTestTypes('', [], []));
     expect(mockOpenrsFetch).toHaveBeenCalledWith(
       expect.stringContaining(
         `${restBaseUrl}/concept?class=Test?v=custom:(display,names:(display),uuid,setMembers:(display,uuid,names:(display),setMembers:(display,uuid,names:(display))))`,
@@ -59,7 +59,7 @@ describe('useTestTypes is configurable', () => {
     );
     await waitFor(() => expect(result.current.isLoading).toBeFalsy());
     expect(result.current.error).toBeFalsy();
-    expect(result.current.orderableConcepts).toEqual([
+    expect(result.current.testTypes).toEqual([
       expect.objectContaining({ conceptUuid: undefined, label: 'Test concept' }),
     ]);
   });
